requirements.txt
attrs==20.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
h11==0.14.0
idna==3.10
jellyfish==0.11.2
lib50==3.0.11
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
submit50==3.1.5
termcolor==1.1.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
